requirements.txt
notion-client==2.2.1
Pillow>=11.0.0
Jinja2>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
PyYAML>=6.0.0
rich>=13.0.0
tenacity>=9.0.0
streamlit>=1.42.0

# Market Analysis
httpx>=0.27.0
pandas>=2.2.0
plotly>=5.24.0
fpdf2>=2.8.0
anthropic>=0.42.0
python-bidi>=0.6.0
arabic-reshaper>=3.0.0
